requirements.txt
# Core dependencies
tqdm>=4.65.0
numpy>=1.24.0

# Optional LLM provider dependencies
# Install based on which providers you want to use:

# OpenAI
openai>=1.12.0
tiktoken>=0.5.0

# Anthropic
anthropic>=0.18.0

# Google Gemini
google-generativeai>=0.3.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
